+++ src/main/java/com/debatetimer/repository/parliamentary/ParliamentaryTableRepository.java
@@ -1,7 +1,7 @@
-package com.debatetimer.repository.parliamentary_debate;
+package com.debatetimer.repository.parliamentary;
 
 import com.debatetimer.domain.member.Member;
-import com.debatetimer.domain.parliamentary_debate.ParliamentaryTable;
+import com.debatetimer.domain.parliamentary.ParliamentaryTable;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
