+++ src/main/java/com/debatetimer/service/member/MemberService.java
@@ -4,9 +4,9 @@ import com.debatetimer.controller.member.dto.MemberCreateRequest;
 import com.debatetimer.controller.member.dto.MemberCreateResponse;
 import com.debatetimer.controller.member.dto.TableResponses;
 import com.debatetimer.domain.member.Member;
-import com.debatetimer.domain.parliamentary_debate.ParliamentaryTable;
+import com.debatetimer.domain.parliamentary.ParliamentaryTable;
 import com.debatetimer.repository.member.MemberRepository;
-import com.debatetimer.repository.parliamentary_debate.ParliamentaryTableRepository;
+import com.debatetimer.repository.parliamentary.ParliamentaryTableRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
