+++ src/test/java/com/debatetimer/BaseServiceTest.java
@@ -1,7 +1,7 @@
 package com.debatetimer;
 
 import com.debatetimer.repository.member.MemberRepository;
-import com.debatetimer.repository.parliamentary_debate.ParliamentaryTableRepository;
+import com.debatetimer.repository.parliamentary.ParliamentaryTableRepository;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
