+++ src/test/java/com/debatetimer/domain/member/MemberTest.java
@@ -1,6 +1,5 @@
-package com.debatetimer.domain;
+package com.debatetimer.domain.member;
 
-import com.debatetimer.domain.member.Member;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
